Group assistant email libs and rename validate_tool module
