+++ front/lib/api/assistant/email/validate_tool_from_email.ts
